+++ HISTORY.md
@@ -2,8 +2,19 @@
 History
 =======
 
+0.1.0 (2022-10-12)
+------------------
+
+* Completely overhauled
+* First stable released with pypi and conda 
+* `plassembler chromosome` added
+* `plassembler custom` added
+* `plassembler mystery` added 
+* `plassembler phage` added
+* `plassembler plasmid` added
+
 
 0.0.1 (2022-10-12)
 ------------------
 
-* First release
+* First release (conda only `conda install -c gbouras dnaapler`)